moved user funct-s from auth.serv to a user.serv
