+++ src/auth/user.service.ts
@@ -1,0 +1,28 @@
+import { UserModel } from './user.model';
+import { Injectable } from '@nestjs/common';
+import { AuthDto } from './dto/auth.dto';
+import { Model } from 'mongoose';
+import { InjectModel } from '@nestjs/mongoose';
+import { genSalt, hash } from 'bcryptjs';
+
+@Injectable()
+export class UserService {
+  constructor(
+    @InjectModel(UserModel.name)
+    private readonly userModel: Model<UserModel>,
+  ) {}
+
+  async createUser(dto: AuthDto) {
+    const salt = await genSalt(10);
+    const newUser = await this.userModel.create({
+      email: dto.login,
+      passwordHash: await hash(dto.password, salt),
+    });
+    return newUser.save();
+  }
+
+  async findUser(email: string) {
+    console.log(email);
+    return this.userModel.findOne({ email }).exec();
+  }
+}
